Fix regressed `HitErrorDisplay` behaviour (and localise binding to meter implementations)
